configured Collection View Cell + Collection View Controller(Pt. 1)
Uncomment Image init + Cell for Row once image is configured

+++ RvBNB/Core data/RvBNB+Convenience.swift
@@ -64,7 +64,7 @@ extension Listing {
           let landDescription = landDescription else {return nil}
         
         
-    return ListingRepresentation(landPhoto: landPhoto, location: location, ownerID: Int(ownerID), pricePerDay: pricePerDay, landDescription: landDescription)
+    return ListingRepresentation(landPhoto: landPhoto, location: location, ownerID: Int64(ownerID), pricePerDay: pricePerDay, landDescription: landDescription)
     }
